refactor(config): move OpenAI resolution into config package
