+++ docs/quickstart.md
@@ -12,13 +12,13 @@ export OPENAI_BASE_URL="http://localhost:8080"
 
 # Cloud (e.g. Google)
 export OPENAI_BASE_URL="https://generativelanguage.googleapis.com/v1beta/openai/"
-export OPENAI_API_KEY="your-api-key"
-export OPENAI_MODEL="your-model-name"
+export OPENAI_API_KEY="your-key"
+export OPENAI_MODEL="your-model"
 ```
+
 > **Windows:** Use your preferred shell's syntax for all environment variables for example `$env:OPENAI_BASE_URL="http://localhost:8080"` in PowerShell.
 
-**2. Launch Late:**
-Late operates within your current working directory. Always launch it from the root of the project you want to work on.
+**2. Launch Late from your project directory:**
 
 ```bash
 cd your-project
@@ -98,8 +98,6 @@ The agent wants to execute bash.
 - **Read-only commands** (`ls`, `cat`, `grep`, etc.) are auto-approved for speed (Note: the listed commands can still require permission if Late deems the agents activity suspicious)
 - **Everything else** requires your explicit `y` / `n`.
 
-> **Note:** On Windows currently, every command will require your explicit `y` / `n` approval. This will be resolved in a future release.
-
 ## Common Flags
 
 | Flag | Description |
@@ -134,17 +132,6 @@ late worktree remove <path>      # Remove a worktree
 
 > **Tip:** Use worktrees when you want Late to work on a feature in the background while you continue working on another branch.
 
-You can also store the same values in Late's app config file. Precedence is: non-empty environment variables > `config.json` > defaults. Empty environment variables fall back to `config.json`.
-
-Windows config path: `%APPDATA%\late\config.json`
-
-```json
-{
-  "openai_base_url": "http://localhost:8080",
-  "openai_api_key": "your-key",
-  "openai_model": "your-model"
-}
-```
 
 ## MCP Integration
 
@@ -162,3 +149,19 @@ Late supports the Model Context Protocol. Add your MCP servers to `~/.config/lat
 ```
 
 MCP tools are automatically available to the agent after connecting.
+
+## Separate Subagent Models
+
+By default, Late uses the same model for both the Lead Architect (orchestrator) and ephemeral workers (subagents). You can mix and match models by setting separate environment variables for subagents:
+
+- **`LATE_SUBAGENT_MODEL`** — The model to use for subagents (e.g., a faster, specialized model).
+- **`LATE_SUBAGENT_BASE_URL`** — (Optional) Different endpoint for subagents. Defaults to `OPENAI_BASE_URL`.
+- **`LATE_SUBAGENT_API_KEY`** — (Optional) Different API key for subagents. Defaults to `OPENAI_API_KEY`.
+
+**Example:** Using a large model for planning and a fast model for execution:
+
+```bash
+export OPENAI_MODEL="o3-mini"
+export LATE_SUBAGENT_MODEL="qwen-32b"
+late
+```
